+++ experimental/CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to experimental packages in this project will be documented 
 
 ### :rocket: (Enhancement)
 
+* feat(instrumentation): Add allowUrls config option to web instrumentation [#4938](https://github.com/open-telemetry/opentelemetry-js/pull/4938) @jairo-mendoza
+
 ### :bug: (Bug Fix)
 
 ### :books: (Refine Doc)